### 2022-08-06 #4656 @arfon
Note to editors – this is a resubmission of https://github.com/openjournals/joss-reviews/issues/4340
