+++ Tutorials/Examples_chemistry/Demo_CASCI_H4.py
@@ -1,8 +1,8 @@
 # ==========================================================================================
 # ==========================================================================================
 # 
-# A code to demonstrate how to implement a FCI method with 
-# the QuantNBody package on the LiH molecule.
+# A code to demonstrate how to implement a CASCI method with 
+# the QuantNBody package on the H4 molecule.
 # The results are compared to the FCI method already implemented in the
 # Psi4 quantum chemistry package.
 #
