### 2022-08-11 0af17c1
Better plot and correction of typos
