### 2022-08-12 1992e4f
Ne function for active space generation
