+++ quantnbody/fermionic/tools.py
@@ -477,7 +477,10 @@ def build_full_mo_1rdm_and_2rdm_for_AS( WFT,
     
     global E_
     global e_
-
+    
+    E_ = np.empty((n_active_mo, n_active_mo), dtype=object)
+    e_ = np.empty((n_active_mo, n_active_mo, n_active_mo, n_active_mo), dtype=object)
+    
     if active_indices is not None:
         for p in range(n_active_mo):
             for q in range(n_active_mo):
@@ -842,7 +845,7 @@ def build_transition_1rdm_spin_free(WFT_A, WFT_B, a_dagger_a):
     transition_one_rdm : spin-free transition 1-RDM
 
     """
-    n_mo = np.shape(a_dagger_a)[0] // 2
+    n_mo = np.shape(a_dagger_a)[0] // 2 
     global E_
     E_ = np.empty((2 * n_mo, 2 * n_mo), dtype=object)
     for p in range(n_mo):
@@ -872,8 +875,7 @@ def build_transition_2rdm_spin_free(WFT_A, WFT_B, a_dagger_a):
     transition_two_rdm : Spin-free transition 2-RDM
 
     """
-    n_mo = np.shape(a_dagger_a)[0] // 2
-    
+    n_mo = np.shape(a_dagger_a)[0] // 2 
     global E_
     E_ = np.empty((2 * n_mo, 2 * n_mo), dtype=object)
     for p in range(n_mo):
@@ -916,11 +918,13 @@ def my_state( slater_determinant, nbody_basis ):
 
     return state
 
+
 def build_projector_active_space( n_elec, 
                                   frozen_indices,
                                   active_indices,
                                   virtual_indices,
-                                  nbody_basis ):
+                                  nbody_basis,
+                                  show_states=False ):
     """
     Build a many-body projector operator including all the many-body configurations
     respecting an active space structure such that :
@@ -944,6 +948,11 @@ def build_projector_active_space( n_elec,
     N_elec_frozen = 2 * len(frozen_indices) 
     state_created_frozen = np.zeros(2*n_mo, dtype=np.int32)
     state_created_active = np.zeros(2*n_mo, dtype=np.int32)
+    
+    if show_states:
+        print()
+        print(' =>> List of states considered in the projector')
+        
     for i in range(n_mo):
         if (i in frozen_indices):
             state_created_frozen[2*i]   = 1
@@ -959,7 +968,9 @@ def build_projector_active_space( n_elec,
                 
             fstate_created = my_state( state, nbody_basis )
             Proj_AS += np.outer(fstate_created, fstate_created)
-            # print(state) 
+            
+            if show_states:  
+               print(state) 
             
     return Proj_AS
 
@@ -1390,6 +1401,7 @@ def scalar_product_different_MO_basis( Psi_A_MOB1,
     return scalar_product
 
 
+
 def transform_psi_MO_basis1_in_MO_basis2( Psi_A_MOB1, 
                                           C_MOB1,
                                           C_MOB2,
@@ -1430,6 +1442,101 @@ def transform_psi_MO_basis1_in_MO_basis2( Psi_A_MOB1,
     
     return Psi_A_MOB2
 
+
+def TEST_transform_psi_MO_basis1_in_MO_basis2( Psi_A_MOB1, 
+                                              C_MOB1,
+                                              C_MOB2,
+                                              nbody_basis, 
+                                              frozen_indices=None):
+    """
+    Transform an intial multi-configurational wavefunction expressed with 
[…]
