### 2022-08-30 #4656 @jarvist
OK @SYalouz , do you have any suggestions for reviewers? I imagine authors of the equivalent Python packages might be well setup to judge the submission from a scientific point of view, and then a Julia experienced programmer from the Julian side.

I note you don't seem to reference QuTIP in your paper, though that seems the established software used by workers in this field? Nor, on the Julia side, Yao.jl.
