### 2022-08-30 #4656 @SYalouz
Dear @jarvist,

Thanks a lot for your message and for considering our work.

Concerning the reviewers, I am thinking of the following list :
- Xavi Monroig-Bonnet ( Openfermion developper, email : bonet@lorentz.leidenuniv.nl )
- Thomas Ayral ( quantum software developper, email: thomas.ayral@atos.net)
- Thomas E. Baker ( Julia experienced programmer in electronic structure, email : thomasedward.baker@york.ac.uk )
- Erik Kjellgren ( electronic structure developper, email : erikkjellgren@live.dk )

Indeed I missed the reference to QuTIP and Yao. Thanks a lot for the suggestions. I will inject the refs in the future corrections of the paper.

Quick question: can I do it right now in the paper.md file, or should I wait for the first round of review ?
