### 2022-08-31 #4656 @jarvist
Dear @SYalouz, Please edit away! It's a continuous process with JOSS. Thank you for your suggestions - I'll now approach some reviewers via email.
