+++ paper.bib
@@ -98,3 +98,42 @@
   pages={1--6},
   year={2015}
 }
+
+
+@article{johansson2012qutip,
+  title={QuTiP: An open-source Python framework for the dynamics of open quantum systems},
+  author={Johansson, J Robert and Nation, Paul D and Nori, Franco},
+  journal={Computer Physics Communications},
+  volume={183},
+  number={8},
+  pages={1760--1772},
+  year={2012},
+  publisher={Elsevier},
+  url={https://doi.org/10.1016/j.cpc.2012.02.021}
+}
+
+@article{johansson2013qutip,
+title = {QuTiP 2: A Python framework for the dynamics of open quantum systems},
+journal = {Computer Physics Communications},
+volume = {184},
+number = {4},
+pages = {1234-1240},
+year = {2013},
+issn = {0010-4655},
+url = {https://doi.org/10.1016/j.cpc.2012.11.019},
+author = {J.R. Johansson and P.D. Nation and Franco Nori},
+keywords = {Open quantum systems, Lindblad, Bloch–Redfield, Floquet–Markov, Master equation, Quantum Monte Carlo, Python},
+abstract = {We present version 2 of QuTiP, the Quantum Toolbox in Python. Compared to the preceding version [J.R. Johansson, P.D. Nation, F. Nori, Comput. Phys. Commun. 183 (2012) 1760.], we have introduced numerous new features, enhanced performance, and made changes in the Application Programming Interface (API) for improved functionality and consistency within the package, as well as increased compatibility with existing conventions used in other scientific software packages for Python. The most significant new features include efficient solvers for arbitrary time-dependent Hamiltonians and collapse operators, support for the Floquet formalism, and new solvers for Bloch–Redfield and Floquet–Markov master equations. Here we introduce these new features, demonstrate their use, and give a summary of the important backward-incompatible API changes introduced in this version.
+Program Summary
+Program title: QuTiP: The Quantum Toolbox in Python Catalog identifier: AEMB_v2_0 Program summary URL:http://cpc.cs.qub.ac.uk/summaries/AEMB_v2_0.html Program obtainable from: CPC Program Library, Queen’s University, Belfast, N. Ireland Licensing provisions: GNU General Public License, version 3 No. of lines in distributed program, including test data, etc.: 33625 No. of bytes in distributed program, including test data, etc.: 410064 Distribution format: tar.gz Programming language: Python. Computer: i386, x86-64. Operating system: Linux, Mac OSX. RAM: 2+ Gigabytes Classification: 7. External routines: NumPy, SciPy, Matplotlib, Cython Catalog identifier of previous version: AEMB_v1_0 Journal reference of previous version: Comput. Phys. Comm. 183 (2012) 1760 Does the new version supercede the previous version?: Yes Nature of problem: Dynamics of open quantum systems Solution method: Numerical solutions to Lindblad, Floquet–Markov, and Bloch–Redfield master equations, as well as the Monte Carlo wave function method. Reasons for new version: Compared to the preceding version we have introduced numerous new features, enhanced performance, and made changes in the Application Programming Interface (API) for improved functionality and consistency within the package, as well as increased compatibility with existing conventions used in other scientific software packages for Python. The most significant new features include efficient solvers for arbitrary time-dependent Hamiltonians and collapse operators, support for the Floquet formalism, and new solvers for Bloch–Redfield and Floquet–Markov master equations. Restrictions: Problems must meet the criteria for using the master equation in Lindblad, Floquet–Markov, or Bloch–Redfield form. Running time: A few seconds up to several tens of hours, depending on size of the underlying
[…]
