+++ paper.bib
@@ -135,5 +135,6 @@ Program title: QuTiP: The Quantum Toolbox in Python Catalog identifier: AEMB_v2_
   volume={4},
   pages={341},
   year={2020},
-  publisher={Verein zur F{\"o}rderung des Open Access Publizierens in den Quantenwissenschaften}
+  publisher={Verein zur F{\"o}rderung des Open Access Publizierens in den Quantenwissenschaften},
+  url={https://doi.org/10.22331/q-2020-10-11-341}
 }
