### 2022-09-14 #4759 @editorialbot
[REVIEW]: QuantNBody: a python package for quantum chemistry and physics to build and manipulate many-body operators and wave functions.
**Submitting author:** @SYalouz (<a href="http://orcid.org/0000-0002-8818-3379">SAAD YALOUZ</a>)
**Repository:** https://github.com/SYalouz/QuantNBody
**Branch with paper.md** (empty if default branch):
**Version:** v 1.0.0
