### 2022-09-14 #4759 @jarvist
Thanks @editorialbot !

OK, @SYalouz, there's some checking cross-references & augmenting your BibTeX suggestions above. You can re-regenerate the paper by asking Editorialbot here, once you've made changes to the repo.
