+++ paper.bib
@@ -41,6 +41,7 @@
   number={1},
   pages={003},
   year={2017},
+  url={10.21468/scipostphys.2.1.003},
   url={https://scipost.org/10.21468/SciPostPhys.2.1.003},
 }
 
@@ -52,7 +53,8 @@
   number={2},
   pages={020},
   year={2019},
-	url={https://scipost.org/10.21468/SciPostPhys.7.2.020},
+  doi={10.21468/SciPostPhys.7.2.020},
+  url={https://scipost.org/10.21468/SciPostPhys.7.2.020},
 }
 
 @article{mcclean2020openfermion,
@@ -136,5 +138,6 @@ Program title: QuTiP: The Quantum Toolbox in Python Catalog identifier: AEMB_v2_
   pages={341},
   year={2020},
   publisher={Verein zur F{\"o}rderung des Open Access Publizierens in den Quantenwissenschaften},
+  url={10.22331/q-2020-10-11-341},
   url={https://doi.org/10.22331/q-2020-10-11-341}
 }
