### 2022-09-14 #4759 @SYalouz
Dear @jarvist,

As recommended, I have added the three missing DOIs in the bib file of the paper. I am going now to regenerate the paper.
