+++ paper.bib
@@ -7,6 +7,7 @@
   pages={3185--3197},
   year={2017},
   publisher={ACS Publications},
+  doi={10.1021/acs.jctc.7b00174},
   url={https://doi.org/10.1021/acs.jctc.7b00174}
 }
 
@@ -19,6 +20,7 @@
   pages={024109},
   year={2020},
   publisher={AIP Publishing LLC},
+  doi = {10.1063/5.0006074},
   url={https://doi.org/10.1063/5.0006074}
 }
 
@@ -41,7 +43,7 @@
   number={1},
   pages={003},
   year={2017},
-  url={10.21468/scipostphys.2.1.003},
+  doi={10.21468/scipostphys.2.1.003},
   url={https://scipost.org/10.21468/SciPostPhys.2.1.003},
 }
 
@@ -66,6 +68,7 @@
   pages={034014},
   year={2020},
   publisher={IOP Publishing},
+  doi = {10.1088/2058-9565/ab8ebc},
   url = {https://doi.org/10.1088/2058-9565/ab8ebc},
 }
 
@@ -111,6 +114,7 @@
   pages={1760--1772},
   year={2012},
   publisher={Elsevier},
+  doi={10.1016/j.cpc.2012.02.021},
   url={https://doi.org/10.1016/j.cpc.2012.02.021}
 }
 
@@ -123,6 +127,7 @@ pages = {1234-1240},
 year = {2013},
 issn = {0010-4655},
 url = {https://doi.org/10.1016/j.cpc.2012.11.019},
+doi={10.1016/j.cpc.2012.11.019},
 author = {J.R. Johansson and P.D. Nation and Franco Nori},
 keywords = {Open quantum systems, Lindblad, Bloch–Redfield, Floquet–Markov, Master equation, Quantum Monte Carlo, Python},
 abstract = {We present version 2 of QuTiP, the Quantum Toolbox in Python. Compared to the preceding version [J.R. Johansson, P.D. Nation, F. Nori, Comput. Phys. Commun. 183 (2012) 1760.], we have introduced numerous new features, enhanced performance, and made changes in the Application Programming Interface (API) for improved functionality and consistency within the package, as well as increased compatibility with existing conventions used in other scientific software packages for Python. The most significant new features include efficient solvers for arbitrary time-dependent Hamiltonians and collapse operators, support for the Floquet formalism, and new solvers for Bloch–Redfield and Floquet–Markov master equations. Here we introduce these new features, demonstrate their use, and give a summary of the important backward-incompatible API changes introduced in this version.
@@ -138,6 +143,6 @@ Program title: QuTiP: The Quantum Toolbox in Python Catalog identifier: AEMB_v2_
   pages={341},
   year={2020},
   publisher={Verein zur F{\"o}rderung des Open Access Publizierens in den Quantenwissenschaften},
-  url={10.22331/q-2020-10-11-341},
+  doi={10.22331/q-2020-10-11-341},
   url={https://doi.org/10.22331/q-2020-10-11-341}
 }
