+++ tools.py
@@ -1,0 +1,2921 @@
+import scipy
+from scipy.optimize import minimize
+import numpy as np
+from itertools import combinations
+from numba import njit, prange 
+import psi4 
+
+E_ = False
+e_ = False
+
+# =============================================================================
+# CORE FUNCTIONS FOR THE BUILDING OF the "A_dagger A" OPERATOR
+# =============================================================================
+
+def build_nbody_basis(n_mo, N_electron, S_z_cleaning=False):
+    """
+    Create a many-body basis formed by a list of slater-determinants
+    (i.e. state encoding occupation numbers in spin-orbitals)
+
+    Parameters
+    ----------
+    n_mo         :  Number of molecular orbitals
+    N_electron   :  Number of electrons
+    S_z_cleaning :  Option if we want to get read of the s_z != 0 states (default is False)
+
+    Returns
+    -------
+    nbody_basis :  List of many-body states (occupation number states) in the basis (occupation number vectors)
+    """
+    # Building the N-electron many-body basis
+    nbody_basis = []
+    for combination in combinations(range(2 * n_mo), N_electron):
+        fock_state = [0] * (2 * n_mo)
+        for index in list(combination):
+            fock_state[index] += 1
+        nbody_basis += [fock_state]
+
+        # In case we want to get rid of states with s_z != 0
+    if S_z_cleaning:
+        nbody_basis_cleaned = nbody_basis.copy()
+        for i in range(np.shape(nbody_basis)[0]):
+            s_z = check_sz(nbody_basis[i])
+            if s_z != 0:
+                nbody_basis_cleaned.remove(nbody_basis[i])
+        nbody_basis = nbody_basis_cleaned
+
+    return np.array(nbody_basis)  # If pybind11 is used it is better to set dtype=np.int8
+
+
+def check_sz(ref_state):
+    """
+    Return the value fo the S_z operator for a unique slater determinant
+    directlyt written as a list of occupation number
+    
+    Parameters
+    ----------
+    ref_state :  Slater determinant (list of occupation numbers)
+
+    Returns
+    -------
+    s_z_slater_determinant : value of S_z for the given slater determinant
+    """
+    s_z_slater_determinant = 0
+    for elem in range(len(ref_state)):
+        if elem % 2 == 0:
+            s_z_slater_determinant += + 1 * ref_state[elem] / 2
+        else:
+            s_z_slater_determinant += - 1 * ref_state[elem] / 2
+
+    return s_z_slater_determinant
+
+
+# numba -> njit version of build_operator_a_dagger_a
+def build_operator_a_dagger_a(nbody_basis, silent=True):
+    """
+    Create a matrix representation of the a_dagger_a operator in the many-body basis
+    
+    Parameters
+    ----------
+    nbody_basis :  List of many-body states (occupation number states) (occupation number states)
+    silent      :  If it is True, function doesn't print anything when it generates a_dagger_a
+    Returns
+    -------
+    a_dagger_a :  Matrix representation of the a_dagger_a operator
+
+    """
+    # Dimensions of problem
+    dim_H = len(nbody_basis)
+    n_mo = len(nbody_basis[0]) // 2
+    mapping_kappa = build_mapping(nbody_basis)
+
+    a_dagger_a = np.zeros((2 * n_mo, 2 * n_mo), dtype=object)
+    for p in range(2 * n_mo):
+        for q in range(p, 2 * n_mo):
+            a_dagger_a[p, q] = scipy.sparse.lil_matrix((dim_H, dim_H))
+            a_dagger_a[q, p] = scipy.sparse.lil_matrix((dim_H, dim_H))
+
+    for MO_q in (range(n_mo)):
[…]
