### 2022-09-18 #4759 @erikkjellgren
# Review checklist for @erikkjellgren

- [ ] I confirm that I have read the [JOSS conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.

- [ ] Repository: Is the source code for this software available at the [QuantNBody](https://github.com/SYalouz/QuantNBody)(https://github.com/DistrictDataLabs/yellowbrick)?
- [ ] License: Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
- [ ] Contribution and authorship: Has the [submitting author](https://github.com/SYalouz) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [ ] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

- [ ] Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
- [ ] A statement of need: Does the paper have a section titled ‘Statement of need’ that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?
- [ ] State of the field: Do the authors describe how this software compares to other commonly-used packages?
- [ ] Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
- [ ] References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax)?
