### 2022-10-03 #4759 @SYalouz
Dear @jarvist,

I contact you regarding citations to the package QuantNbody.

There are actually two papers from my group which have been realised using QuantNbody.  One is already on arxiv, (see: https://arxiv.org/abs/2209.10302) and submitted to a journal, the second one will appear on arxiv by the end of the week (and submitted to a journal also this week).

I am wondering: May I inject explicit references to both these papers in the current manuscript ? This in order to demonstrate the interest of the package in current research projects.

Thanks in advance for your reply
