### 2022-10-11 #4759 @SYalouz
Dear @jarvist ,

I come back to you to mention that the second paper from my group is now released on Arxiv (see https://arxiv.org/abs/2210.02325) where the package QuantNBody was employed. This paper has been submitted to a journal and it is now under review process.

I come back to you with the same questions as before: May I inject explicit references to both these papers in the current manuscript ? This in order to demonstrate the interest of the package in current research projects.

Thanks in advance for your reply
