### 2022-10-11 #4759 @jarvist
I see no problem with this, as long as there is a suitable scholarly discussion to put them in context.
