### 2022-10-11 #4759 @SYalouz
Ok, I have injected the two articles referencing the QuantNBody package, I will now generate the latest PDF version for the manuscript.
