### 2022-10-16 #4759 @erikkjellgren
Comments for review points that might not be fulfilled:
I hope I have done it in an acceptable format @jarvist

__Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.__
There is working installation instruction on the main GitHub page of the package.
The instructions instruct to install Psi4 separately from QuantNbody, instead of having it as a direct dependency.
This installation setup is not what is described as ideal. But it is easy enough.

__Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).__
In the GitHub repo of the package, there is a folder called Tutorials.
This folder contains some notebooks with descriptions and code, which is a nice way to show how to use the code.
The folder also contains some example_* folders with just plain code. A README.md file in the Tutorial folder would be appreciated to guide a user to what to actually look at.

__Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?__
I failed to find any documentation other than README.md in the root directory of the GitHub repo.

__Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?__
I failed to find any testing setup for the codebase.

__Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support__
I failed to find any community guidelines for the package.

__State of the field: Do the authors describe how this software compares to other commonly-used packages?__
The authors make a list of software that is comparable in 'philosophy'. It is a little unclear if QuantNbody is more "theoretician-friendly" than the already existing packages.

__Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?__
Line 1: "python" should be with capitalized "P".

Line 11: "hamiltonians" should be with capitalized "H".

Line 33: It should be "SciPy" instead of "Scipy".

Line 39: The sentence seems fragmented, perhaps something is missing in place of the ":"

Line 42:  "hamiltonians" should be with capitalized "H".

Line 57: "hubbard" should be with capitalized "H".

Line 58: The sentence starting here is a little unclear.
             If I understand correctly, the message is that the the Hamiltonians in Eq (1) and Eq (2) are 'programmed' using the single-body hopping oprator?
			 Or is something more techinical happening?

Line 62: "python" should be with capitalized "P".

Line 64:  "hamiltonians" should be with capitalized "H".

Line 65: H2 should be in roman instead of italic, https://iupac.org/wp-content/uploads/2019/05/IUPAC-GB3-2012-2ndPrinting-PDFsearchable.pdf

Figure 1: Atoms should be in roman. (H2 in figure text should be in roman)
