### 2022-10-17 #4759 @jarvist
Dear Erik, @erikkjellgren
Thank you very much! These look quite detailed, and the suggested corrections well described.

@SYalouz , if you could check through & rebut or apply changes to Erik's suggestions, and address any weaknesses / issues. Please don't lose heart! JOSS is very much about working openly to improve the codes, paper and documentation to a suitable standard.
