### 2022-10-23 SYalouz/QuantNBody#2 @wcwitt
JOSS Review - Installation
For the [JOSS review](https://github.com/openjournals/joss-reviews/issues/4759).

> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

I managed to install the package itself using the recommended `git clone` and `pip`. The README mentions installing `psi4` first, and I managed that with `conda`.

However, some additional packages were required to complete the tutorials, including some nonstandard ones like `tqdm` and `pyscf`. Please make a fresh install of everything and track all the extra packages required for the tutorials, adding these to the documentation.
