### 2022-10-23 SYalouz/QuantNBody#3 @wcwitt
JOSS Review - Examples
Part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/4759).

> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)

There are a number of examples in the `Tutorials` directory, either as Jupyter notebooks or as scripts. However, the top-level documentation does not contain a full summary of the examples, and it takes some effort to understand what the the scripts (in particular) are meant to do. Please list all the examples in the documentation, with a short summary for each one.

I have completed the Jupyter notebook tutorials, which provide a good amount of context and explanation. However, they have some minor issues with spelling, grammar, and overall tidyness. Please make sure they are polished to the same level as the paper.

There are some additional files in the `Tutorials` directory whose purpose is not obvious to me (namely, `timer.dat`, `graph.png`, and  `OLD tuto`). Please make sure this directory is tidy so that new users can navigate it easily.
