### 2022-10-23 SYalouz/QuantNBody#4 @wcwitt
JOSS Review - Documentation
Part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/4759).

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

There do seem to be doc strings in much of the source code, but that is the only documentation I found. Please use ReadTheDocs, Sphinx, or something equivalent to create web-based documentation that organizes these doc strings into full documentation. Ideally, the examples would also appear as part of such documentation.
