### 2022-10-23 SYalouz/QuantNBody#5 @wcwitt
JOSS Review - Tests
Part of the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/4759).

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I found very few tests and nothing automated. Please increase the test coverage and set up something something automated using GitHub actions or similar.
