### 2022-10-25 a7df6ed
first corrections of the tutorials
