### 2022-10-25 ea81712
Delete index.rst
