+++ docs/conf.py
@@ -24,7 +24,7 @@ author = 'Saad Yalouz'
 html_theme = "sphinx_rtd_theme"
 html_theme_path = ["_themes", ]
 html_static_path = ['_static']
-html_logo = "_static/logo.png"
+html_logo = "logo.png"
 
 
 # -- Options for LaTeX output ---------------------------------------------
