+++ docs/source/install.rst
@@ -1,0 +1,18 @@
+Installation and Running
+========================
+
+Prerequisite Packages
+__________
+In order to run QuantNBody you will need to have next packages:
+
+    scipy, numpy, numba
+
+Installation
+___________
+To install the package you can either download the package from `PyPI <https://pypi.org/>`_:
+
+    pip install quantnbody
+
+or clone the repository
+
+    git clone in developer mode ?? TODO
