### 2022-10-31 4ab633d
Adding first automated tests to the package
