### 2022-10-31 b773e0c
Adding first function in rst as an example and editing docstring of function to be better transformed to webpage
