+++ docs/source/api.rst
@@ -1,0 +1,6 @@
+QuanNBody API
+=============
+
+.. currentmodule:: quantnbody.fermionic.tools
+
+.. autofunction:: build_nbody_basis
