+++ docs/source/index.rst
@@ -1,0 +1,35 @@
+.. QuantNBody documentation master file, created by
+   sphinx-quickstart on Sat Oct 29 14:31:41 2022.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+QuantNBody
+==========
+
+.. toctree::
+   :maxdepth: 2
+   :caption: How to use
+
+   install.rst
+   examples.rst
+   issues.rst
+   structure.rst
+
+.. toctree::
+   :maxdepth: 2
+   :caption: API
+
+   api.rst
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Data structures
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
