### 2022-10-31 cc5a4e6
Adding autodoc function descriptions
