+++ docs/source/conf.py
@@ -18,7 +18,8 @@ release = '1.0.1'
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = ['sphinx.ext.autodoc']
+extensions = ['sphinx.ext.autodoc',
+              "sphinx.ext.napoleon"]
 
 templates_path = ['_templates']
 exclude_patterns = []
