+++ docs/source/api.rst
@@ -1,6 +1,212 @@
 QuanNBody API
 =============
 
+fermionic.tools
+_______________
+
 .. currentmodule:: quantnbody.fermionic.tools
 
+In this subsection, there are written functions for fermionic particles.
+
 .. autofunction:: build_nbody_basis
+
+.. autofunction:: check_sz
+
+.. autofunction:: build_operator_a_dagger_a
+
+.. autofunction:: build_mapping
+
+.. autofunction:: make_integer_out_of_bit_vector
+
+.. autofunction:: new_state_after_sq_fermi_op
+
+.. autofunction:: build_final_state_ad_a
+
+.. autofunction:: build_hamiltonian_quantum_chemistry
+
+.. autofunction:: build_hamiltonian_fermi_hubbard
+
+.. autofunction:: build_penalty_orbital_occupancy
+
+.. autofunction:: build_E_and_e_operators
+
+.. autofunction:: build_full_mo_1rdm_and_2rdm_for_AS
+
+.. autofunction:: build_1rdm_alpha
+
+.. autofunction:: build_1rdm_beta
+
+.. autofunction:: build_1rdm_spin_free
+
+.. autofunction:: build_2rdm_fh_on_site_repulsion
+
+.. autofunction:: build_2rdm_fh_dipolar_interactions
+
+.. autofunction:: build_2rdm_spin_free
+
+.. autofunction:: build_1rdm_and_2rdm_spin_free
+
+.. autofunction:: build_hybrid_1rdm_alpha_beta
+
+.. autofunction:: build_transition_1rdm_alpha
+
+.. autofunction:: build_transition_1rdm_beta
+
+.. autofunction:: build_transition_1rdm_spin_free
+
+.. autofunction:: build_transition_2rdm_spin_free
+
+.. autofunction:: my_state
+
+.. autofunction:: build_projector_active_space
+
+.. autofunction:: fh_get_active_space_integrals
+
+.. autofunction:: fh_get_active_space_integrals_with_V
+
+.. autofunction:: qc_get_active_space_integrals
+
+.. autofunction:: build_s2_sz_splus_operator
+
+.. autofunction:: build_s2_local
+
+.. autofunction:: build_sAsB_coupling
+
+.. autofunction:: build_spin_subspaces
+
+.. autofunction:: get_info_from_psi4
+
+.. autofunction:: weight_det
+
+.. autofunction:: scalar_product_different_MO_basis
+
+.. autofunction:: transform_psi_MO_basis1_in_MO_basis2
+
+.. autofunction:: TEST_transform_psi_MO_basis1_in_MO_basis2
+
+.. autofunction:: scalar_product_different_MO_basis_with_frozen_orbitals
+
+.. autofunction:: visualize_wft
+
+.. autofunction:: get_ket_in_atomic_orbitals
+
+.. autofunction:: compute_energy_with_rdm
+
+.. autofunction:: prepare_vector_k_orbital_rotation_fwith_active_space
+
+.. autofunction:: transform_vec_to_skewmatrix_with_active_space
+
+.. autofunction:: energy_cost_function_orbital_optimization
+
+.. autofunction:: brute_force_orbital_optimization
+
+.. autofunction:: filter_h_g_orb
+
+.. autofunction:: sa_build_mo_hessian_and_gradient
+
+.. autofunction:: sa_build_mo_hessian_and_gradient_no_active_space
+
+.. autofunction:: build_mo_gradient
+
+.. autofunction:: orbital_optimisation_newtonraphson
+
+.. autofunction:: orbital_optimisation_newtonraphson_no_active_space
+
+.. autofunction:: transform_vec_to_skewmatrix
+
+.. autofunction:: build_generalized_fock_matrix
+
+.. autofunction:: build_generalized_fock_matrix_active_space_adapted
+
[…]
