### 2022-11-01 c2003f0
Raise exception if type_of_op is not either a or a^
