### 2022-11-01 2dd378e
#1 part of correcting docstrings to NumPy style
