+++ quantnbody/fermionic/tools.py
@@ -17,12 +17,19 @@ def build_nbody_basis(n_mo, N_electron, S_z_cleaning=False):
     Create a many-body basis formed by a list of slater-determinants
     (i.e. state encoding occupation numbers in spin-orbitals)
 
-    :param int n_mo:  Number of molecular orbitals
-    :param int N_electron:  Number of electrons
-    :param bool S_z_cleaning:  Option if we want to get read of the s_z != 0 states (default is False)
-    :returns:  ``nbody_basis``:  List of many-body states (occupation number states) in the basis
-        (occupation number vectors)
-    :rtype: NumPy array
+    Parameters
+    ----------
+    n_mo : int
+        Number of molecular orbitals
+    N_electron :  int
+        Number of electrons
+    S_z_cleaning : bool, default=False
+        Option if we want to get read of the s_z != 0 states (default is False)
+
+    Returns
+    -------
+    ``nbody_basis``
+        List of many-body states (occupation number states) in the basis (occupation number vectors)
     """
     # Building the N-electron many-body basis
     nbody_basis = []
@@ -47,15 +54,17 @@ def build_nbody_basis(n_mo, N_electron, S_z_cleaning=False):
 def check_sz(ref_state):
     """
     Return the value fo the S_z operator for a unique slater determinant
-    directlyt written as a list of occupation number
+    directly written as a list of occupation number
     
     Parameters
     ----------
-    ref_state :  Slater determinant (list of occupation numbers)
+    ref_state :  Iterable[int]
+        Slater determinant (list of occupation numbers)
 
     Returns
     -------
-    s_z_slater_determinant : value of S_z for the given slater determinant
+    s_z_slater_determinant
+        value of S_z for the given slater determinant
     """
     s_z_slater_determinant = 0
     for elem in range(len(ref_state)):
@@ -74,11 +83,15 @@ def build_operator_a_dagger_a(nbody_basis, silent=True):
     
     Parameters
     ----------
-    nbody_basis :  List of many-body states (occupation number states) (occupation number states)
-    silent      :  If it is True, function doesn't print anything when it generates a_dagger_a
+    nbody_basis : Iterable[Iterable[int]]
+        List of many-body states (occupation number states)
+    silent : bool, default=True
+        If it is True, function doesn't print anything when it generates a_dagger_a
+
     Returns
     -------
-    a_dagger_a :  Matrix representation of the a_dagger_a operator
+    a_dagger_a
+        Matrix representation of the a_dagger_a operator
 
     """
     # Dimensions of problem
@@ -148,11 +161,13 @@ def build_mapping(nbody_basis):
 
     Parameters
     ----------
-    nbody_basis :  List of many-body state written in terms of occupation numbers
+    nbody_basis :  Iterable[Iterable[int]]
+        List of many-body state written in terms of occupation numbers
  
     Returns
     -------
-    mapping_kappa : List of unique values associated to each kappa
+    mapping_kappa : Iterable[int]
+        List of unique values associated to each kappa
     """
     num_digits = np.shape(nbody_basis)[1]
     dim_H = np.shape(nbody_basis)[0]
@@ -175,11 +190,13 @@ def make_integer_out_of_bit_vector(ref_state):
 
     Parameters
     ----------
-    ref_state : Reference slater determinant to turn out into an integer
+    ref_state : Iterable[int]
+        Reference slater determinant to turn out into an integer
 
     Returns
     -------
-    number : unique integer referring to the slater determinant
+    int
+        Number of unique integer referring to the slater determinant.
     """
     number = 0
     for digit in range(len(ref_state)):
[…]
