### 2022-11-01 8e9e8e8
- second part of correcting docstrings to NumPy style - started to add examples in docstring
