+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'QuantNBody:  a python package for quantum chemistry and physics to build and manipulate many-body operators and wave functions.'
+title: 'QuantNBody:  a Python package for quantum chemistry and physics to build and manipulate many-body operators and wave functions.'
 tags:
   - Python
   - quantum physics and chemistry
@@ -23,7 +23,7 @@ bibliography: paper.bib
 
 # Summary
 
-The 'QuantNBody' package is a Python toolkit for quantum chemists/physicists interested in methods development to study quantum many-body problems ranging from electronic structure to condensed matter theory. It provides a quick and easy way to build matrix representations of bosonic and fermionic quantum many-body operators (*e.g.* hamiltonians, spin or excitation operators) and get access to quantities/objects of interest (*e.g.* energies, reduced density matrices, many-body wave functions). The code includes various native functions and it is flexible enough to help users in building their own numerical tools to implement new methods. 
+The 'QuantNBody' package is a Python toolkit for quantum chemists/physicists interested in methods development to study quantum many-body problems ranging from electronic structure to condensed matter theory. It provides a quick and easy way to build matrix representations of bosonic and fermionic quantum many-body operators (*e.g.* Hamiltonians, spin or excitation operators) and get access to quantities/objects of interest (*e.g.* energies, reduced density matrices, many-body wave functions). The code includes various native functions and it is flexible enough to help users in building their own numerical tools to implement new methods. 
 
 # Statement of need
 
@@ -33,7 +33,7 @@ The 'QuantNBody' code was designed to answer this problem: it is a theoretician-
 
 # Framework of the package
 
-The QuantNbody package employs the Scipy Package [@2020SciPy-NMeth] for the creation of sparse matrices representation of many-body operators. The Numba package [@lam2015numba] is also used to accelerate calculation when possible. The framework of the package lies in two fundamental ingredients. The first one is the creation of a reference many-body vector basis (based on a total number of quantum particles and modes/orbitals to fill) in which second quantization operators can be represented. The second ingredient consists in creating a general tool that can help build any particle-number conserving many-body operator : the single-body hopping operators $a^\dagger_p a_q$.  Once these two ingredients have been created, the user can employ various pre-built functions in order to (i) construct different types of many-body operators (*e.g.* hamiltonians, spin and excitation operators), (ii) manipulate/visualize quantum many-body states. All the native functions have been thought to facilitate calculations for new users and young students. Beyond this, the QuantNBody package has been also designed to provide flexibility to experimented users to develop their own tools to implement/test their own methods.
[…]
