### 2022-11-02 6c0484d
Added example for fermi hubbard model (without reference calculation)
