+++ paper.md
@@ -29,7 +29,9 @@ The 'QuantNBody' package is a Python toolkit for quantum chemists/physicists int
 
 The manipulation of many-body operators in the language of second quantization is a crucial step for accessing the properties of model or *ab initio* systems in quantum chemistry and physics.  From a numerical point of view, this requires developing codes that can build matrix representations of quantum operators in a given quantum many-body basis (*e.g.* Hamiltonians, spin or excitation operators).  In a great majority of cases, this aspect is kept as a ''blackbox'' in codes to spare the users from cumbersome numerical parts and to facilitate the use of already implemented methods. Nevertheless, this type of implementation can appear as a real obstacle to overcome for researchers in need of reliable numerical tools to quickly develop and test new methodologies based on second quantization algebra. 
 
-The 'QuantNBody' code was designed to answer this problem: it is a theoretician-friendly package which provides an efficient and simple-to-use numerical toolkit to help create and manipulate operators and wavefunctions related to quantum many-body systems (with either fermionic or bosonic particles). Its philosophy is thus comparable to the one of other packages such as Quimb [@gray2018quimb], Quspin [@weinberg2017quspin;@weinberg2019quspin], OpenFermion [@mcclean2020openfermion], QuTIP [@johansson2012qutip;@johansson2013qutip] or Yao [@luo2020yao].
+The 'QuantNBody' code was designed to answer this problem: it is a theoretician-friendly package which provides an efficient and simple-to-use numerical toolkit to help create and manipulate operators and wavefunctions related to quantum many-body systems (with either fermionic or bosonic particles). The specificity of the package  is to allow a straightforward implementation of second quantization operators into numerical objects (*i.e.* matrices). This in order to accelerate and facilitate the conversion of analytical developments into compact and easy to write numerical Python codes. This use-friendly philosophy goes along the lines of other recent packages dedicated to the simulations of many-body systems such as Quimb [@gray2018quimb], Quspin [@weinberg2017quspin;@weinberg2019quspin], OpenFermion [@mcclean2020openfermion], QuTIP [@johansson2012qutip;@johansson2013qutip] or Yao [@luo2020yao]. 
+To our knowledge, QuantNBody is the only Python toolbox proposing this user-friendly approach for the encoding and manipulation of second quantization objects.
+
 
 # Framework of the package
 
