### 2022-11-02 df02c12
adding new sections in the readme
