+++ README.md
@@ -51,7 +51,7 @@ python -m pip install -e .
 Different examples and tutorials are furnished in the tutorial/ repository in python or under the form of Jupyter notebooks.  
 
 
- ## How to contribute:
+ ## How to contribute
 
 
 We'd love to accept your contributions and patches to QuantNBody. There are a few small guidelines you need to follow.  
@@ -59,7 +59,7 @@ We'd love to accept your contributions and patches to QuantNBody. There are a fe
 All submissions require review. We use GitHub pull requests for this purpose. Consult GitHub Help for more information on using pull requests. Furthermore, please make sure your new code comes with documentation.
 
 
- ## Support: 
+ ## Support
  
 If you are having issues, please let us know by posting the issue on our Github issue tracker.
   
