+++ Tutorials/readme.md
@@ -4,13 +4,13 @@ In this folder we gather a series of tutorials/examples of how to use the QuantN
 
 ## Notebooks tutorials
 
-In order to learn the how to domesticate the package, we encourage the new users to follow the jupyter noteboks tutorials in the following order:
+In order to learn how to domesticate the package, we encourage the new users to read the jupyter notebooks tutorials in the following order:
 
 - *Tuto_FIRST_STEP.ipynb:*
 
 This tutorial explains the basics of the QuantNBody package. Focusing on fermionic systems, 
-we explain how the encoding a many-body basis is realized in practice in the code. We also detail how we encode the single-body hopping operators.
-The latter being a central tools for the creation of any particle number conserving operator in the code. 
+we explain how the encoding of a many-body basis (spin orbitals occupied by electrons) is realized in practice in the code.
+We also detail how we encode the single-body hopping operators. The latter being a central tool for the creation of any particle number conserving operators in the code. 
 
 - *Tuto_PLAYING_WITH_STATES.ipynb:*
 
@@ -28,4 +28,8 @@ For those interesting in bosonic systems, we also describe here equivalent featu
 
 ## Examples of scripts
 
-We also provide a series of python scripts illustrating different types of many-body calculations implemented fully with the QuantNBody packages. The folder are named
+We also provide a series of python scripts illustrating different types of many-body calculations. All these methods are implemented from scratch with the QuantNBody packages and range from FCI, CASCI or even CASSCF calculation (for fermionic systems). The folder are named according to the types of systems used as a study case:
+
+- *Examples_BoseHubbard:* script for FCI calculation on small-sized Bose-Hubbard systems.
+- *Examples_FermiHubbard:* scripts for CAS-CI and FCI calculation on small-sized Fermi-Hubbard systems.
+- *Examples_chemistry:* scripts for CAS-CI, SA-CASSCF and FCI calculation on small-sized molecular systems.
