+++ Tutorials/readme.md
@@ -6,24 +6,16 @@ In this folder we gather a series of tutorials/examples of how to use the QuantN
 
 In order to learn how to domesticate the package, we encourage the new users to read the jupyter notebooks tutorials in the following order:
 
-- *Tuto_FIRST_STEP.ipynb:*
-
-This tutorial explains the basics of the QuantNBody package. Focusing on fermionic systems, 
+- *Tuto_FIRST_STEP.ipynb:* This tutorial explains the basics of the QuantNBody package. Focusing on fermionic systems, 
 we explain how the encoding of a many-body basis (spin orbitals occupied by electrons) is realized in practice in the code.
 We also detail how we encode the single-body hopping operators. The latter being a central tool for the creation of any particle number conserving operators in the code. 
 
-- *Tuto_PLAYING_WITH_STATES.ipynb:*
+- *Tuto_PLAYING_WITH_STATES.ipynb:* This second tutorial illustrates how to easily manipulate multi-body states with the help of different illustrative examples. 
 
-This second tutorial illustrates how to easily manipulate multi-body states with the help of different illustrative examples. 
-
-- *Tuto_SPIN_AND_AB_INITIO_HAMILTONIANS.ipynb:*
-
-This third tutorials focuses on the construction of different spin operators (e.g. $\hat{S}_2$) and ab initio electronic structure Hamiltonians.
+- *Tuto_SPIN_AND_AB_INITIO_HAMILTONIANS.ipynb:* This third tutorials focuses on the construction of different spin operators (e.g. $\hat{S}_2$) and ab initio electronic structure Hamiltonians.
 We will show how easily we can build these operators and use them.
 
-- *Tuto_BOSE_HUBBARD.ipynb:*
-
-For those interesting in bosonic systems, we also describe here equivalent features/functions to build operators, see/manipulate wavefunctions.
+- *Tuto_BOSE_HUBBARD.ipynb:* For those interesting in bosonic systems, we also describe here equivalent features/functions to build operators, see/manipulate wavefunctions.
 
 
 ## Examples of scripts
