### 2022-11-02 6621302
Deleting unecessary tqdm use
