### 2022-11-03 fe485fc
Added logo in the docs
