+++ docs/source/conf.py
@@ -19,16 +19,21 @@ release = '1.0.1'
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
 extensions = ['sphinx.ext.autodoc',
-              "sphinx.ext.napoleon"]
+              'sphinx.ext.napoleon',
+              'sphinx_rtd_theme',
+              'nbsphinx',]
 
 templates_path = ['_templates']
 exclude_patterns = []
 
 html_static_path = ['_static']
-html_logo = "_static/logo.png"
+html_logo = "_static/logo2.png"
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
-html_theme = 'alabaster'
+html_theme = 'sphinx_rtd_theme'
 html_static_path = ['_static']
+html_theme_options = {
+    'navigation_depth': 4,
+}
