+++ docs/source/issues.rst
@@ -1,6 +1,0 @@
-Issues
-======
-
-In this section known problems with the code can be found
-
-TODO
