+++ docs/source/Tuto_SPIN_AND_AB_INITIO_HAMILTONIANS.rst
@@ -1,0 +1,365 @@
+*QuantNbody* tutorials : electronic structure Hamiltonian and spin operators
+============================================================================
+
+**Dr. Saad Yalouz - Laboratoire de Chimie Quantique de Strasbourg,
+France - July 2022**
+
+In this QuantNBody tutorial, we will focus on the construction of spin
+operators like :math:`S_2` and :math:`S_z` and ab initio electronic
+structure Hamiltonians :math:`H`. We will show how easily we can build
+these operators and use them.
+
+Let us first import our package and some additional libraries
+
+.. code:: ipython3
+
+    import quantnbody as qnb
+    import numpy as np
+    import scipy
+    import psi4
+    import math
+
+Electronic structure Hamiltonian
+--------------------------------
+
+The ab initio electronic structure Hamiltonian describes how a
+collection of :math:`N_e` electrons rearrange around a set of positively
+charged nuclei in a molecule. This operator is defined like:
+
+.. math::  H = \sum_{p,q} h_{pq} {E}_{pq} + \frac{1}{2} \sum_{pqrs} g_{pqrs} {e}_{pqrs} + \mathcal{E}_{nuc.\ rep.}
+
+where we have the one- and two-electron integrals :math:`h_{pq}` and
+:math:`g_{pqrs}` (and the nuclear repulsion energy
+:math:`\mathcal{E}_{nuc.\ rep.}`) associated to the one- and
+two-electron spin-free excitation operators defined like
+
+.. math::  {E}_{pq} = \sum_{\sigma=\uparrow,\downarrow} a^\dagger_{p,\sigma}a_{q,\sigma}
+
+and
+
+.. math::  e_{pqrs} = E_{pq}E_{rs} - \delta_{rs} E_{ps} .
+
+Here again, we see that all these operators can be defined in terms of
+the :math:`a_{p,\sigma}^\dagger a_{q,\tau}` operators.
+
+Below, we show how to create such a Hamiltonian for the :math:`H_4`
+molecule in the STO-3G basis, and how to (easily) compute a ground state
+potential energy surface. In this case, we need to get access to the
+electronic integrals of the system. We use for this Psi4 (but note that
+Pyscf could have been used too !)
+
+First we build the two essential ingredients: the many-body basis and :math:`a^\dagger a` operators
+
+.. code:: ipython3
+
+    N_elec = N_MO = 4 # Number of MOs and electrons in the system
+
+    nbody_basis = qnb.fermionic.tools.build_nbody_basis( N_MO, N_elec )
+    a_dagger_a  = qnb.fermionic.tools.build_operator_a_dagger_a( nbody_basis )
+
+Then we can build the PES of the molecule with a loop
+
+.. code:: ipython3
+
+    def RUN_FCI_PSI4( string_geo, basisset, num_roots ):
+        '''
+        A function to run a FCI method with the Psi4 package
+        '''
+        psi4.geometry( string_geo )
+        psi4.set_options({  'basis'       : basisset,
+                            'num_roots'   : num_roots,
+                            'S'           : 0.0,
+                            'DIAG_METHOD' : 'SEM' })
+        fci, fcici_wfn = psi4.energy('fci',return_wfn=True)
+
+        E_FCI = []
+        for root in range(num_roots):
+            E_FCI += [ psi4.variable('CI ROOT {} TOTAL ENERGY'.format(root)) ]
+
+        return E_FCI
+
+    E_HF  = []
+    E_FCI = []
+    E_FCI_QNB = []
+
+    N_elec   = N_MO = 4
+    basisset = 'sto-3g'
[…]
