+++ docs/source/install.rst
@@ -1,18 +1,24 @@
 Installation and Running
 ========================
 
-Prerequisite Packages
-_____________________
-In order to run QuantNBody you will need to have next packages:
+.. note::
 
-    scipy, numpy, numba, psi4
+  To run the **QuantNBody** package, you will need first to install the next four packages:
 
-Installation
-____________
-To install the package you can either download the package from `PyPI <https://pypi.org/>`_:
+  #. **SciPy** : for installing, see https://scipy.org/install/
+  #. **NumPy** : for installing, see https://numpy.org/install/
+  #. **Numba** : for installing, see https://numba.readthedocs.io/en/stable/user/installing.html
+  #. **Psi4**  : for installing, see https://anaconda.org/psi4/psi4, or https://psicode.org/psi4manual/1.2.1/conda.html
 
-    pip install quantnbody
 
-or clone the repository
+To install the package you can download the package from `PyPI <https://pypi.org/>`_:
 
-    git clone in developer mode ?? TODO
+    >>> pip install quantnbody
+
+Another alternative is to clone the repository from Github and install it in developper mode:
+
+  >>> git clone https://github.com/SYalouz/QuantNBody.git
+  >>> cd QuantNBody
+  >>> python -m pip install -e .
+
+And dont forget to :code:`git pull` to always get the latest version of the package !
