+++ docs/source/Tuto_FIRST_STEP.rst
@@ -1,0 +1,420 @@
+*QuantNbody* tutorials : first steps with the package
+=====================================================
+
+**Dr. Saad Yalouz - Laboratoire de Chimie Quantique de Strasbourg,
+France - July 2022**
+
+Philosophy of the package
+-------------------------
+
+The philosophy of the **QuantNBody** package is to facilitate the
+implementation and manipulation of quantum many-body systems composed of
+electrons or bosons. To achieve this goal, the package has been designed
+to provide a fast and easy way to construct many-body operators and wave
+functions in a given many-body basis. In this way, it becomes possible
+to access quantities/objects of interest in a few lines of python to
+speed up and facilitate method development.
+
+To do this, the package works with two fundamental ingredients.
+
+A) The first is the **creation of a many-body basis** (based on a total
+   number of quantum particles and modes/orbitals to be filled) in which
+   each operator can be represented.
+
+B) The second is the creation of the set of :math:`a^\dagger a`\ \*\*
+   hopping operators that are needed to construct any many-body operator
+   that conserves the number of particles.
+
+Once these two ingredients are created, the user can use predefined
+functions to construct different types of many-body operators
+(e.g. Hamiltonians, spin operators), and manipulate/view many-body
+quantum states. Note that the QuantNBody package has also been designed
+to provide flexibility for users to also create their own operators and
+functions based on the tools provided.
+
+**Nota Bene:** For the sake of simplicity, we will concentrate in these
+tutorials on fermionic systems.
+
+Let us first import the package !
+---------------------------------
+
+.. code:: ipython3
+
+    import quantnbody as qnb   # <==== General import
+    import numpy as np
+
+Building a many-body basis
+--------------------------
+
+To build a many-body basis for a fermionic system, the QuantNBody
+package generates a list of many-body states which describe the
+repartition of :math:`N_{elec}` electrons in :math:`2N_{MO}`
+spin-orbitals. These states are numerically referenced by a list of
+kappa indices such that
+
+.. math::
+
+
+   \Big\lbrace |\kappa \rangle \Big\rbrace_{\textstyle \kappa=1}^{\textstyle \dim_H}
+
+The dimension :math:`\dim_H` of the many-body basis depends on the
+number of electron :math:`N_{elec}` and spatial orbital :math:`N_{MO}`
+via a binomial law such that
+
+.. math:: \dim_H = \binom{2N_{MO}}{N_{elec}}
+
+**A little example with** :math:`N_{MO}=N_{elec}=2`: in this case, we
+should have **6 many-body states.**
+
+.. code:: ipython3
+
+    N_MO = N_elec = 2 # We define the numebr of MO adn electrons
+
+    nbody_basis = qnb.fermionic.tools.build_nbody_basis( N_MO, N_elec ) # Building the nbody_basis
+
+    print('Shape  of the kappa states')
+    for s in range(len(nbody_basis)):
+        print('| kappa={} >'.format(s), '=', nbody_basis[s])
+
+ 
+.. code:: none
+
+    Shape  of the kappa states
+    | kappa=0 > = [1 1 0 0]
[…]
