+++ docs/source/remarks_structure.rst
@@ -4,8 +4,8 @@ Note on the structure of the package
 .. note::
   The package **QuantNBody** is stuctured in 2 subpackages :code:`quantnbody.fermionic` and :code:`quantnbody.bosonic` which are respectively dedicated to fermionic and bosonic systems.
   Each subpackage contains a list of functions (encapsulated in a local :code:`tools`
-  subfolder) one can use for creating/manipulating many-body operators/wavefunctions.
-  In practice, to access all the functions, one only needs to import the following repositories from quantnbody:
+  module) one can use for creating/manipulating many-body operators/wavefunctions.
+  In practice, to access all the functions, one only needs to import the following modules from quantnbody:
 
 
   For fermionic systems
