### 2022-11-04 1527587
modif for the examples
