+++ docs/source/install.rst
@@ -11,14 +11,15 @@ Installation and Running
   #. **Psi4**  : for installing, see https://anaconda.org/psi4/psi4, or https://psicode.org/psi4manual/1.2.1/conda.html
 
 
-To install the package you can download the package from `PyPI <https://pypi.org/>`_:
+An easy way to install the QuantNBody package is by using "pip" in your terminal as follows:
 
     >>> pip install quantnbody
 
-Another alternative is to clone the repository from Github and install it in developper mode:
+Another alternative is to clone the repository from our `Github page <https://github.com/SYalouz/QuantNBody>`_  and install it in developper mode.
+For this, write the following lines in your terminal:
 
   >>> git clone https://github.com/SYalouz/QuantNBody.git
   >>> cd QuantNBody
   >>> python -m pip install -e .
 
-And dont forget to :code:`git pull` to always get the latest version of the package !
+And dont forget to always :code:`git pull` if you want to get the latest version of the package !
