### 2022-11-04 8200148
working on the functions
