+++ docs/source/requirements.txt
@@ -2,3 +2,7 @@ nbsphinx
 sphinx
 sphinxemoji
 readthedocs-sphinx-ext
+psi4
+numpy
+scipy
+numba
