+++ docs/source/conf.py
@@ -23,6 +23,8 @@ extensions = ['sphinx.ext.autodoc',
               'sphinx_rtd_theme',
               'nbsphinx',
               'sphinxemoji.sphinxemoji']
+ 
+napoleon_numpy_docstring = True 
 
 templates_path = ['_templates']
 exclude_patterns = []
