+++ quantnbody/bosonic/tools.py
@@ -1,10 +1,10 @@
 import scipy
 import numpy as np
 from itertools import combinations_with_replacement
-from numba import njit, prange 
+from numba import njit, prange
 import scipy.sparse
-from tqdm import tqdm 
- 
+from tqdm import tqdm
+
 # =============================================================================
 # CORE FUNCTIONS FOR THE BUILDING OF the "A_dagger A" OPERATOR
 # =============================================================================
@@ -12,16 +12,19 @@ from tqdm import tqdm
 def build_nbody_basis( n_mode, n_boson ):
     """
     Create a many-body basis formed by a list of fock-state with a conserved total
-    number of bosons 
+    number of bosons
 
     Parameters
     ----------
-    n_mode    :  Number of modes in total 
-    N_boson   :  Number of bosons in total 
+    n_mode    : int
+        Number of modes in total
+    N_boson   :  int
+        Number of bosons in total
 
     Returns
     -------
-    nbody_basis :  List of many-body states (occupation number states)  
+    nbody_basis : array
+        List of many-body states (occupation number states)
     """
     # Building the N-electron many-body basis
     nbody_basis = []
@@ -30,9 +33,9 @@ def build_nbody_basis( n_mode, n_boson ):
         for index in list(combination):
             fock_state[ index ] += 1
         nbody_basis += [ fock_state ]
-        
+
     return np.array(nbody_basis)  # If pybind11 is used it is better to set dtype=np.int8
- 
+
 
 @njit
 def build_mapping( nbodybasis ):
@@ -42,24 +45,26 @@ def build_mapping( nbodybasis ):
 
     Parameters
     ----------
-    nbody_basis :  Many-body basis
+    nbody_basis : array
+        Many-body basis
 
     Returns
     -------
-    mapping_kappa : List of unique values associated to each kappa
+    mapping_kappa : array
+        List of unique values associated to each kappa
     """
     n_mode  = np.shape(nbodybasis)[1]
     n_boson = np.sum(nbodybasis[0])
     max_number = n_boson * 10**(( n_mode - 1 )*2) + 1
-    dim_H = np.shape(nbodybasis)[0] 
+    dim_H = np.shape(nbodybasis)[0]
     mapping_kappa = np.zeros( max_number, dtype=np.int32 )
     for kappa in range(dim_H):
         ref_state = nbodybasis[ kappa ]
-        number = 0 
-        for index_mode in range(n_mode):  
+        number = 0
+        for index_mode in range(n_mode):
             number += ref_state[index_mode] * 10**(( n_mode - index_mode - 1 )*2)
         mapping_kappa[ int(np.round(number)) ] = kappa
-        
+
     return mapping_kappa
 
 
@@ -67,17 +72,17 @@ def build_mapping( nbodybasis ):
 # def build_final_state_ad_a(ref_state, p, q, mapping_kappa):
 #     state_one, coeff1 = new_state_after_sq_boson_op('a',  q, ref_state)
 #     state_two, coeff2 = new_state_after_sq_boson_op('a^', p, state_one)
-#     kappa_ = mapping_kappa[ make_number_out_of_vector(state_two) ]  
- 
-#     return kappa_, coeff1, coeff2
- 
+#     kappa_ = mapping_kappa[ make_number_out_of_vector(state_two) ]
 
-# OLDER and MORE "BRUTE FORCE" FUNCTION KEPT FOR DEBUGGING 
+#     return kappa_, coeff1, coeff2
+
+
+# OLDER and MORE "BRUTE FORCE" FUNCTION KEPT FOR DEBUGGING
 # @njit
 def build_final_state_ad_a(ref_state, p, q, nbodybasis):
     state_one, coeff1 = new_state_after_sq_boson_op('a',  q, ref_state)
[…]
