+++ quantnbody/fermionic/tools.py
@@ -2,8 +2,8 @@ import scipy
 from scipy.optimize import minimize
 import numpy as np
 from itertools import combinations
-from numba import njit, prange 
-import psi4 
+from numba import njit, prange
+import psi4
 
 E_ = False
 e_ = False
@@ -14,8 +14,9 @@ e_ = False
 
 def build_nbody_basis(n_mo, N_electron, S_z_cleaning=False):
     """
-    Create a many-body basis formed by a list of slater-determinants
-    (i.e. state encoding occupation numbers in spin-orbitals)
+    Create a many-body basis as a list of slater-determinants. Here, these states are
+    occupation numbersvectors taking the form of bitstrings -- *e.g.* |1100⟩ --
+    describing how the electrons occupy the spin-orbitals.
 
     Parameters
     ----------
@@ -28,8 +29,8 @@ def build_nbody_basis(n_mo, N_electron, S_z_cleaning=False):
 
     Returns
     -------
-    ``nbody_basis``
-        List of many-body states (occupation number states) in the basis (occupation number vectors)
+    nbody_basis : array
+        List of many-body states (occupation number vectors). The col
 
     Examples
     ________
@@ -64,16 +65,16 @@ def build_nbody_basis(n_mo, N_electron, S_z_cleaning=False):
 def check_sz(ref_state):
     """
     Return the value of the S_z operator for a unique slater determinant
-    directly written as a list of occupation number
-    
+    (written as a list of occupation number)
+
     Parameters
     ----------
-    ref_state :  Iterable[int]
+    ref_state :  vector
         Slater determinant (list of occupation numbers)
 
     Returns
     -------
-    s_z_slater_determinant
+    s_z_slater_determinant : float
         value of S_z for the given slater determinant
 
     Examples
@@ -98,24 +99,24 @@ def check_sz(ref_state):
 def build_operator_a_dagger_a(nbody_basis, silent=True):
     """
     Create a matrix representation of the a_dagger_a operator in the many-body basis
-    
+
     Parameters
     ----------
-    nbody_basis : Iterable[Iterable[int]]
+    nbody_basis : array
         List of many-body states (occupation number states)
     silent : bool, default=True
         If it is True, function doesn't print anything when it generates a_dagger_a
 
     Returns
     -------
-    a_dagger_a
-        Matrix representation of the a_dagger_a operator
+    a_dagger_a : array
+        Matrix representation of the a_dagger_a operators
 
     Examples
     ________
     >>> nbody_basis = nbody_basis(2, 2)
-    >>> a_dagger_a = build_operator_a_dagger_a(nbody_basis, True) 
-    >>> a_dagger_a[0,0] # Get access to the operator counting the electron in the first spinorbital 
+    >>> a_dagger_a = build_operator_a_dagger_a(nbody_basis, True)
+    >>> a_dagger_a[0,0] # Get access to the operator counting the electron in the first spinorbital
     """
     # Dimensions of problem
     dim_H = len(nbody_basis)
@@ -186,7 +187,7 @@ def build_mapping(nbody_basis):
     ----------
     nbody_basis :  Iterable[Iterable[int]]
         List of many-body state written in terms of occupation numbers
- 
+
     Returns
     -------
     mapping_kappa : Iterable[int]
@@ -242,9 +243,9 @@ def make_integer_out_of_bit_vector(ref_state):
 @njit
 def new_state_after_sq_fermi_op(type_of_op, index_mode, ref_fock_state):
     """
-    Create the final state generated by the application of a second quantization 
+    Create the final state generated by the application of a second quantization
[…]
