+++ docs/source/references.rst
@@ -1,0 +1,7 @@
+Works using QuantNBody
+======================= 
+
+
+* Yalouz, Saad, Sajanthan Sekaran, and Matthieu Saubanère. `"Quantum embedding of multi-orbital fragments using the Block-Householder-transformation." arXiv preprint arXiv:2209.10302 (2022) <https://arxiv.org/abs/2209.10302>`_
+
+* Roseiro, Pablo, David JR Brook, Nadia Ben Amor, Vincent Robert, and Saad Yalouz. `"Excited state spinmerism in high-field Fe (II)-verdazyl molecular complex: versatile local spins for quantum information." arXiv preprint arXiv:2210.02325 (2022) <https://arxiv.org/abs/2206.03978>`_

[…]
