### 2022-11-05 573f8d5
Cleaning more the fermionic functions
