+++ quantnbody/fermionic/tools.py
@@ -30,7 +30,7 @@ def build_nbody_basis(n_mo, N_electron, S_z_cleaning=False):
     Returns
     -------
     nbody_basis : array
-        List of many-body states (occupation number vectors). The col
+        List of many-body states (occupation number vectors).
 
     Examples
     ________
@@ -356,7 +356,7 @@ def build_hamiltonian_quantum_chemistry(h_,
     S_2_target : float, default=None
         Value of the S_2 mean value we want to target (default is None)
     penalty : float, default=100
-        Value of the penalty term for state not respecting the spin symmetry (default is 100).
+        Value of the penalty term to penalize the states that do not respect the spin symmetry (default is 100).
 
     Returns
     -------
@@ -431,14 +431,14 @@ def build_hamiltonian_fermi_hubbard(h_,
     S_2_target : float, default=None
         Value of the S_2 mean value we want to target (default is None)
     penalty : float, default=100
-        Value of the penalty term for state not respecting the spin symmetry (default is 100).
+        Value of the penalty term to penalize the states that do not respect the spin symmetry (default is 100).
     v_term : array, default=None
         dipolar interactions.
 
     Returns
     -------
     H_fermi_hubbard : array
-        Matrix representation of the Fermi-Hubbard Hamiltonian int eh many-body basis
+        Matrix representation of the Fermi-Hubbard Hamiltonian in the many-body basis
 
     """
     # # Dimension of the problem
@@ -788,7 +788,7 @@ def build_2rdm_fh_on_site_repulsion(WFT, a_dagger_a, mask=None):
 
 def build_2rdm_fh_dipolar_interactions(WFT, a_dagger_a, mask=None):
     """
-    Create a spin-free 2 RDM out of a given Fermi Hubbard wave function for dipolar interaction operator
+    Create a spin-free 2 RDM for a given Fermi Hubbard wave function for dipolar interaction operator
     it corresponds to <psi|(a^+_i↑ a_j↑ + a^+_i↓ a_j↓)(a^+_k↑ a_l↑ + a^+_k↓ a_l↓)|psi>
 
     Parameters
@@ -918,7 +918,7 @@ def build_1rdm_and_2rdm_spin_free(WFT, a_dagger_a):
 
 def build_hybrid_1rdm_alpha_beta(WFT, a_dagger_a):
     """
-    Create a hybrid alpha-beta 1 RDM out of a given wave function
+    Create a hybrid alpha-beta 1 RDM for a given wave function
     (Note : alpha for the lines, and beta for the columns)
 
     Parameters
@@ -945,7 +945,7 @@ def build_hybrid_1rdm_alpha_beta(WFT, a_dagger_a):
 
 def build_transition_1rdm_alpha(WFT_A, WFT_B, a_dagger_a):
     """
-    Create a spin-alpha transition 1 RDM out of a given wave function
+    Create a spin-alpha transition 1 RDM for a given wave function
 
     Parameters
     ----------
@@ -973,7 +973,7 @@ def build_transition_1rdm_alpha(WFT_A, WFT_B, a_dagger_a):
 
 def build_transition_1rdm_beta(WFT_A, WFT_B, a_dagger_a):
     """
-    Create a spin-beta transition 1 RDM out of a given wave function
+    Create a spin-beta transition 1 RDM for a given wave function
 
     Parameters
     ----------
@@ -1170,7 +1170,7 @@ def fh_get_active_space_integrals ( h_MO,
                                     frozen_indices=None,
                                     active_indices=None ):
     """
-    Restricts a Fermi-Hubbard system at a spatial orbital level to an active space
+    Restricts a Fermi-Hubbard system at a spatial orbital level to an active space.
[…]
