+++ docs/source/conf.py
@@ -36,6 +36,7 @@ html_logo = "_static/logo2.png"
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
 html_theme = 'sphinx_rtd_theme'
+# html_theme = 'press'
 html_static_path = ['_static']
 html_theme_options = {
     'navigation_depth': 4,
