+++ docs/source/references.rst
@@ -1,7 +1,14 @@
 Works using QuantNBody
-======================= 
+=======================
 
+We mention below a (non-exhaustive) list of papers/works where the **QuantNBody** package was used to produce results or develop new methods.
 
-* Yalouz, Saad, Sajanthan Sekaran, and Matthieu Saubanère. `"Quantum embedding of multi-orbital fragments using the Block-Householder-transformation." arXiv preprint arXiv:2209.10302 (2022) <https://arxiv.org/abs/2209.10302>`_
+.. note ::
+  Please, feel free to contact us if you want your work to be mentionned here. We will be happy to add
+  your ref to this list |:smiley:|
 
-* Roseiro, Pablo, David JR Brook, Nadia Ben Amor, Vincent Robert, and Saad Yalouz. `"Excited state spinmerism in high-field Fe (II)-verdazyl molecular complex: versatile local spins for quantum information." arXiv preprint arXiv:2210.02325 (2022) <https://arxiv.org/abs/2206.03978>`_
+**List of papers/works:**
+
+  * Yalouz, Saad, Sajanthan Sekaran, and Matthieu Saubanère. `"Quantum embedding of multi-orbital fragments using the Block-Householder-transformation." arXiv preprint arXiv:2209.10302 (2022) <https://arxiv.org/abs/2209.10302>`_
+
+  * Roseiro, Pablo, David JR Brook, Nadia Ben Amor, Vincent Robert, and Saad Yalouz. `"Excited state spinmerism in high-field Fe (II)-verdazyl molecular complex: versatile local spins for quantum information." arXiv preprint arXiv:2210.02325 (2022) <https://arxiv.org/abs/2206.03978>`_
