### 2022-11-06 6a39010
Removing old rst file that is not actually used for the docs
