### 2022-11-06 6860932
- Changing docstings to double-quoted strings - adding additional new line in the end for sphinx autodocs
