+++ quantnbody/fermionic/tools.py
@@ -41,6 +41,7 @@ def build_nbody_basis(n_mo, N_electron, S_z_cleaning=False):
            [0, 1, 1, 0],
            [0, 1, 0, 1],
            [0, 0, 1, 1]])
+
     """
     # Building the N-electron many-body basis
     nbody_basis = []
@@ -84,6 +85,7 @@ def check_sz(ref_state):
 
     >>> check_sz([1, 1, 0, 0])  # in reference state there is doubly occupied first MO
     0.0
+
     """
     s_z_slater_determinant = 0
     for elem in range(len(ref_state)):
@@ -117,6 +119,7 @@ def build_operator_a_dagger_a(nbody_basis, silent=True):
     >>> nbody_basis = nbody_basis(2, 2)
     >>> a_dagger_a = build_operator_a_dagger_a(nbody_basis, True)
     >>> a_dagger_a[0,0] # Get access to the operator counting the electron in the first spinorbital
+
     """
     # Dimensions of problem
     dim_H = len(nbody_basis)
@@ -198,6 +201,7 @@ def build_mapping(nbody_basis):
     >>> nbody_basis = build_nbody_basis(2, 2)
     >>> build_mapping(nbody_basis)
     array([0, 0, 0, 5, 0, 4, 3, 0, 0, 2, 1, 0, 0, 0, 0, 0])
+
     """
     num_digits = np.shape(nbody_basis)[1]
     dim_H = np.shape(nbody_basis)[0]
@@ -232,6 +236,7 @@ def make_integer_out_of_bit_vector(ref_state):
     ________
     >>> make_integer_out_of_bit_vector(np.array([1, 1, 0, 0]))
     12
+
     """
     number = 0
     for digit in range(len(ref_state)):
@@ -273,6 +278,7 @@ def new_state_after_sq_fermi_op(type_of_op, index_mode, ref_fock_state):
     (array([0, 1, 0, 0]), 1.0)
     >>> qnb.new_state_after_sq_fermi_op("a^", 3, np.array([1, 1, 0, 0])) # creation of electron
     (array([1, 1, 0, 1]), 1.0)
+
     """
     new_fock_state = ref_fock_state.copy()
     coeff_phase = (-1.) ** np.sum(ref_fock_state[0:index_mode])
@@ -318,6 +324,7 @@ def build_final_state_ad_a(ref_state, p, q, mapping_kappa):
     >>> mapping = build_mapping(nbody_basis)
     >>> build_final_state_ad_a(np.array([1, 1, 0, 0]), 2, 1, mapping)  # exciting electron from spin MO 1 to spin MO 2
     (1, -1.0, -1.0)
+
     """
     state_one, p1 = new_state_after_sq_fermi_op('a', q, ref_state)
     state_two, p2 = new_state_after_sq_fermi_op('a^', p, state_one)
@@ -515,6 +522,7 @@ def build_penalty_orbital_occupancy( a_dagger_a, occupancy_target):
              [-0.4, -1.6,  0. ,  0. ,  0. ,  4. ]]),
      array([], dtype=int64),
      array([0, 1, 2, 3, 4, 5], dtype=int64))
+
     """
     dim_H = a_dagger_a[0,0].shape[0]
     n_mo  = np.shape(a_dagger_a)[0] // 2
@@ -552,6 +560,7 @@ def build_E_and_e_operators( a_dagger_a, n_mo ):
     >>> nbody_basis = qnb.build_nbody_basis(2, 2)
     >>> a_dagger_a = qnb.build_operator_a_dagger_a(nbody_basis)
     >>> build_E_and_e_operators(a_dagger_a, 2)
+
     """
     E_ = np.empty((n_mo, n_mo), dtype=object)
[…]
