+++ docs/source/Tuto_BOSE_HUBBARD.rst
@@ -1,5 +1,5 @@
 *Tuto 4:* the Bose-Hubbard system
-===================
+=================================
 
 **Dr. Saad Yalouz - Laboratoire de Chimie Quantique de Strasbourg,
 France - July 2022**
