### 2022-11-06 aea008a
updating readme and docs
