+++ README.md
@@ -15,8 +15,9 @@ It provides a quick and easy way to build many-body operators and wavefunctions 
 
  ## Documentation of the package
 
-[See the documentation here !](https://quantnbody.readthedocs.io/en/latest/) 
-
+<div style="text-align: center;">
+[>>>>> See the documentation here ! <<<<<](https://quantnbody.readthedocs.io/en/latest/) 
+</div>
 
 We provide below a non-exhaustive list of the various possibilities offered by the package:
 
