+++ README.md
@@ -19,7 +19,7 @@ It provides a quick and easy way to build many-body operators and wavefunctions 
   
 [>>>>> See the documentation here ! <<<<<](https://quantnbody.readthedocs.io/en/latest/) 
                                                                                         
-<\div>
+</div>
 
 We provide below a non-exhaustive list of the various possibilities offered by the package:
 
