+++ README.md
@@ -15,7 +15,7 @@ It provides a quick and easy way to build many-body operators and wavefunctions 
 
  ## Documentation of the package
 
-<div style="text-align: center;">
+<div align="center">
   
 [>>>>> See the documentation here ! <<<<<](https://quantnbody.readthedocs.io/en/latest/) 
                                                                                         
