### 2022-11-06 014b485
Modifiying docstrings to comply with the sphinx, adding forgotten function argument
