### 2022-11-06 2926c53
Restructuring Householder transformation funtions so now they are rendered better in html form
