### 2022-11-07 dbd8325
new tests and correction doc
