+++ docs/source/index.rst
@@ -55,5 +55,5 @@ This tool can be also of a great help for pedagogical purpose and to help illust
 Indices and tables
 ==================
 
-* :ref:`genindex` 
+* :ref:`genindex`
 * :ref:`search`
