### 2022-11-07 89ebd52
unittest version of new_tests.py
