+++ testing/new_tests_unittest.py
@@ -1,0 +1,201 @@
+import unittest
+import numpy as np
+import psi4
+import scipy
+import quantnbody as qnb
+
+
+class MyTestCase(unittest.TestCase):
+    def test_FERMION_BASICS(self):
+        # ========================================================|
+        # Parameters for the simulation
+        nelec_active = 4  # Number of active electrons in the Active-Space
+        active_indices = [i for i in range(0, 3)]
+        n_mo_active = len(active_indices)
+
+        # Building the Many-body basis
+        nbody_basis = qnb.fermionic.tools.build_nbody_basis(len(active_indices), nelec_active)
+
+        # Building the matrix representation of the adagger_a operator in the many-body basis
+        a_dagger_a = qnb.fermionic.tools.build_operator_a_dagger_a(nbody_basis)
+
+        self.assertTrue(np.allclose(nbody_basis, np.load('nbodybasis_fermion_test.npy'), atol=1e-08),
+                        'Error in many-body basis generation for fermionic systems')
+        test_a_dagger_a = np.load('a_dagger_a_fermion_test.npy', allow_pickle=True)
+
+        for mode in range(2 * n_mo_active):
+            for mode_ in range(2 * n_mo_active):
+                self.assertTrue(np.allclose(a_dagger_a[mode, mode_].A, test_a_dagger_a[mode, mode_].A, atol=1e-08),
+                                'Error in a_dagger_a generation for fermionic systems')
+
+        print(nbody_basis, np.load('nbodybasis_fermion_test.npy'))
+
+    def test_FERMION_MODEL_HAMILTONIAN(self):
+        # ========================================================|
+        # Parameters for the simulation
+        nelec_active = 4  # Number of active electrons in the Active-Space
+        frozen_indices = [i for i in range(0)]
+        active_indices = [i for i in range(0, 3)]
+        virtual_indices = [i for i in range(4, 4)]
+
+        n_mo = len(frozen_indices) + len(active_indices) + len(virtual_indices)
+
+        # Building the Many-body basis
+        nbody_basis = qnb.fermionic.tools.build_nbody_basis(len(active_indices), nelec_active)
+
+        # Building the matrix representation of the adagger_a operator in the many-body basis
+        a_dagger_a = qnb.fermionic.tools.build_operator_a_dagger_a(nbody_basis)
+
+        # Building the matrix representation of several interesting spin operators in the many-body basis
+        S_2, s_z, s_plus = qnb.fermionic.tools.build_s2_sz_splus_operator(a_dagger_a)
+
+        U = 6
+
+        # Hopping terms
+        h_MO = np.zeros((n_mo, n_mo))
+        for site in range(n_mo):
+            for site_ in range(site, n_mo - 1):
+                h_MO[site, site_] = h_MO[site_, site] = -1
+        h_MO[0, n_mo - 1] = h_MO[n_mo - 1, 0] = -1
+
+        # MO energies
+        for site in range(n_mo):
+            h_MO[site, site] = - site
+
+        U_MO = np.zeros((n_mo, n_mo, n_mo, n_mo))
+        for site in range(n_mo):
+            U_MO[site, site, site, site] = U
+
+        # Preparing the active space integrals and the associated core contribution
+        E_core, h_, U_ = qnb.fermionic.tools.fh_get_active_space_integrals(h_MO,
+                                                                           U_MO,
+                                                                           frozen_indices=frozen_indices,
+                                                                           active_indices=active_indices)
+
+        # Building the matrix representation of the Hamiltonian operators
+        H = qnb.fermionic.tools.build_hamiltonian_fermi_hubbard(h_,
+                                                                U_,
+                                                                nbody_basis,
[…]
