### 2022-11-07 055258e
Forgot to replace 2 assert statements
