+++ testing/new_tests_unittest.py
@@ -28,8 +28,6 @@ class MyTestCase(unittest.TestCase):
                 self.assertTrue(np.allclose(a_dagger_a[mode, mode_].A, test_a_dagger_a[mode, mode_].A, atol=1e-08),
                                 'Error in a_dagger_a generation for fermionic systems')
 
-        print(nbody_basis, np.load('nbodybasis_fermion_test.npy'))
-
     def test_FERMION_MODEL_HAMILTONIAN(self):
         # ========================================================|
         # Parameters for the simulation
@@ -166,7 +164,7 @@ class MyTestCase(unittest.TestCase):
         for mode in range(n_mode):
             for mode_ in range(n_mode):
                 self.assertTrue(np.allclose(a_dagger_a[mode, mode_].A,
-                                   test_a_dagger_a[mode, mode_].A), 'Error in groundstate energy estimation')
+                                            test_a_dagger_a[mode, mode_].A), 'Error in groundstate energy estimation')
 
     def test_BOSON_MODEL_HAMILTONIAN(self):
         n_mode = 4
@@ -191,7 +189,7 @@ class MyTestCase(unittest.TestCase):
                                                              nbodybasis,
                                                              a_dagger_a)
         eig_en, eig_vec = scipy.linalg.eigh(H.A)
-        self.assertTrue(np.allclose(H.A, np.load('ham_boson_test.npy',allow_pickle=True)),
+        self.assertTrue(np.allclose(H.A, np.load('ham_boson_test.npy', allow_pickle=True)),
                         'Error in Hamiltonian generation for bosonic systems')
         self.assertAlmostEqual(eig_en[0], -49.276698175251404, 8, 'Error in groundstate energy estimation')
         self.assertAlmostEqual(eig_en[1], -49.27144736991592, 8, 'Error in first excited state energy estimation')
