+++ docs/source/requirements.txt
@@ -6,3 +6,4 @@ numpy
 scipy
 numba
 tqdm
+ipython
