+++ Tutorials/Examples_chemistry/Demo_SACASSCF_LiH.py
@@ -144,7 +144,7 @@ for r in (list_r):
 
     E_core, h_, g_ = qnb.fermionic.tools.qc_get_active_space_integrals(h_MO,
                                                                        g_MO,
-                                                                       occupied_indices=frozen_indices,
+                                                                       frozen_indices=frozen_indices,
                                                                        active_indices=active_indices)
 
     # %%
@@ -217,7 +217,7 @@ for r in (list_r):
         # Building the orbital-optimized Frozen-core Hamiltonian
         E_core, h_, g_ = qnb.fermionic.tools.qc_get_active_space_integrals(h_MO,
                                                                            g_MO,
-                                                                           occupied_indices=frozen_indices,
+                                                                           frozen_indices=frozen_indices,
                                                                            active_indices=active_indices)
 
         # Building the matrix representation of the Hamiltonian operators 
