+++ README.md
@@ -53,11 +53,11 @@ cd QuantNBody
 python -m pip install -e .
 ```
  
- Note that you'll need to install the Psi4 package before installing QuantNBody. For this we redirect the user to the following link:
+Note that you'll need to install the Psi4 package before installing QuantNBody. For this we redirect the user to the following link:
  
  - Psi4 installations : [Using conda](https://anaconda.org/psi4/psi4), see also the [following link](https://psicode.org/psi4manual/1.2.1/conda.html)
 
-Once the package is fully installed, you can run some tests to check if everything was correclty installed. For this, go the the testing folder and run the following line in your terminal:
+Once the package is fully installed, you can run some tests to check if everything was correctly done. For this, go the the [testing folder](https://github.com/SYalouz/QuantNBody/tree/main/testing) and run the following line in your terminal:
 ```
 python TESTS.py
 ```
