### 2022-11-07 d4cd49d
cleaning functions
