+++ paper.md
@@ -31,7 +31,7 @@ The manipulation of many-body operators in the language of second quantization i
 
 The 'QuantNBody' code was designed to answer this problem: it is a theoretician-friendly package which provides an efficient and simple-to-use numerical toolkit to help create and manipulate operators and wavefunctions related to quantum many-body systems (with either fermionic or bosonic particles). It allows straightforward implementations of second quantization into numerical objects (*i.e.* matrices) in order to facilitate the conversion of analytical developments into compact and easy to write numerical Python codes. 
 
-The user-friendly philosophy of QuantNBody regarding second quantization encoding can be compared (to some extent) to the FermiLib module originally developped for OpenFermion [@mcclean2020openfermion] (a quantum computing oriented package). In practice, the advantage of QuantNBody lies in its particle number conserving encoding  which differs from what is done in Openfermion (where the full Fockspace is encoded). As a result, QuantNBody allows the realization of larger scale calculations (up to 8 electrons in 8 spatial molecular orbitals in a few seconds). This new toolkit opens new perspectives for  many-body simulations in Python and then complete the panel of already existing theoretician-friendly packages dedicated to many-body systems such as Quimb [@gray2018quimb], Quspin [@weinberg2017quspin;@weinberg2019quspin], QuTIP [@johansson2012qutip;@johansson2013qutip], Yao [@luo2020yao] and OpenFermion [@mcclean2020openfermion].  
+The user-friendly philosophy of QuantNBody regarding second quantization encoding can be compared (to some extent) to the FermiLib module originally developped for OpenFermion [@mcclean2020openfermion] (a quantum computing oriented package). In practice, the advantage of QuantNBody lies in its particle number conserving encoding  which differs from what is done in Openfermion (where the full Fockspace is encoded). As a result, QuantNBody allows the realization of larger scale calculations (up to 8 electrons in 8 spatial molecular orbitals in a few seconds). This new toolkit opens new perspectives for  many-body simulations in Python and then completes the panel of already existing theoretician-friendly packages dedicated to many-body systems such as Quimb [@gray2018quimb], Quspin [@weinberg2017quspin;@weinberg2019quspin], QuTIP [@johansson2012qutip;@johansson2013qutip], Yao [@luo2020yao] and OpenFermion [@mcclean2020openfermion].  
 
 
 # Framework of the package
