### 2022-11-08 57a0480
corrections in the examples scripts for Psi4
