### 2022-11-08 #4759 @SYalouz
Dear @erikkjellgren, see below my answers to all your comments !

Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution. There is working installation instruction on the main GitHub page of the package. The instructions instruct to install Psi4 separately from QuantNbody, instead of having it as a direct dependency. This installation setup is not what is described as ideal. But it is easy enough.

**_Our answer:_**  We agree with you: it is not necessarily ideal but it works.  We have been fighting with the installation of this package.
We essentially came up with the simple idea of directly asking the user to install it.

Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems). In the GitHub repo of the package, there is a folder called Tutorials. This folder contains some notebooks with descriptions and code, which is a nice way to show how to use the code. The folder also contains some example_* folders with just plain code. A README.md file in the Tutorial folder would be appreciated to guide a user to what to actually look at.

**_Our answer:_**  To guide new users, we added a README.md file in the “example” folder that details all the tutorials and examples.
Furthermore, we want to mention that the tutorials have been also included in the new documentation.

Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)? I failed to find any documentation other than README.md in the root directory of the GitHub repo.

**_Our answer:_**  It is an important point that was also raised by the second reviewer.  Thus, we generated a complete documentation with “readthedoc” that can be accessed here: https://quantnbody.readthedocs.io/en/latest/.

Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified? I failed to find any testing setup for the codebase.

**_Our answer:_** We created a dedicated folder (called “testing”) to run tests for both fermionic and bosonic systems. We detailed in the main README.md file how to run these tests.

Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support I failed to find any community guidelines for the package.

**_Our answer:_**  We added more information to the main README.md file at the root of the project. Now people will know how to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.

=== State of the field: Do the authors describe how this software compares to other commonly-used packages? === The authors make a list of software that is comparable in 'philosophy'. It is a little unclear if QuantNbody is more "theoretician-friendly" than the already existing packages.

**_Our answer:_** We understand that the text here lacked clarity. We have therefore extended the description of the specificity of the package by highlighting the fact that QuantNbody is among the few packages that can offer a numerical encoding of second quantization operators for fermionic and bosonic systems.

Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)? === Line 1: "python" should be with capitalized "P".
=== Line 11: "hamiltonians" should be with capitalized "H".
=== Line 33: It should be "SciPy" instead of "Scipy".
=== Line 39: The sentence seems fragmented, perhaps something is missing in place of the ":"
=== Line 42: "hamiltonians" should be with capitalized "H".
=== Line 57: "hubbard" should be with capitalized "H".
== Line 58: The sentence starting here is a little unclear If I understand correctly, the message is that the the Hamiltonians in Eq (1) and Eq (2) are 'programmed' using the single-body hopping oprator? Or is something more techinical happening?
=== Line 62: "python" should be with capitalized "P".
=== Line 64: "hamiltonians" should be with capitalized "H".
Line 65: H2 should be in roman instead of italic, https://iupac.org/wp-content/uploads/2019/05/IUPAC-GB3-2012-2ndPrinting-PDFsearchable.pdf
Figure 1: Atoms should be in roman. (H2 in figure caption should be in roman)

**_Our answer:_**  We implemented all the corrections suggested in the paper.
