### 2022-11-08 SYalouz/QuantNBody#4 @SYalouz
It is an important point that was also raised by the second reviewer. Thus, we generated a complete documentation with “readthedoc” that can be accessed here: https://quantnbody.readthedocs.io/en/latest/.
