### 2022-11-08 SYalouz/QuantNBody#3 @SYalouz
To guide new users, we added a README.md file in the “example” folder that details all the tutorials and examples.  We have also checked grammar and spelling. Furthermore, we want to mention that the tutorials have been also included in the new documentation.
