### 2022-11-08 SYalouz/QuantNBody#5 @SYalouz
We created a dedicated folder (called “testing”) to run tests for both fermionic and bosonic systems. We detailed in the main README.md file how to run these tests.
