### 2022-11-09 f31c28b
Changing setup.py for the PyPi upload
