+++ docs/source/requirements.txt
@@ -5,5 +5,4 @@ readthedocs-sphinx-ext
 numpy
 scipy
 numba
-tqdm
 ipython
