### 2022-11-16 d02b37b
update works mentionning QNB
