+++ docs/source/references.rst
@@ -9,8 +9,8 @@ We mention below a (non-exhaustive) list of papers/works where the **QuantNBody*
 
 **List of papers/works:**
 
-  * Yalouz, Saad, Sajanthan Sekaran, and Matthieu Saubanère. `"Quantum embedding of multi-orbital fragments using the Block-Householder-transformation." arXiv preprint arXiv:2209.10302 (2022) <https://arxiv.org/abs/2209.10302>`_
+  * Saad Yalouz, Sajanthan Sekaran, and Matthieu Saubanère. `"Quantum embedding of multi-orbital fragments using the Block-Householder-transformation." arXiv preprint arXiv:2209.10302 (2022) <https://arxiv.org/abs/2209.10302>`_
 
-  * Roseiro, Pablo, David JR Brook, Nadia Ben Amor, Vincent Robert, and Saad Yalouz. `"Excited state spinmerism in high-field Fe (II)-verdazyl molecular complex: versatile local spins for quantum information." arXiv preprint arXiv:2210.02325 (2022) <https://arxiv.org/abs/2206.03978>`_
+  * Pablo Roseiro, David JR Brook, Nadia Ben Amor, Vincent Robert, and Saad Yalouz. `"Excited state spinmerism in high-field Fe (II)-verdazyl molecular complex: versatile local spins for quantum information." arXiv preprint arXiv:2210.02325 (2022) <https://arxiv.org/abs/2206.03978>`_
 
-  * Saad Yalouz and  Vincent Robert. `"Orthogonally Constrained Orbital Optimization: assessing changes of optimal orbitals for orthogonal multi-reference states." arXiv preprint arXiv:2211.08329 (2022) <https://doi.org/10.48550/arXiv.2211.08329>`_
+  * Saad Yalouz and Vincent Robert. `"Orthogonally Constrained Orbital Optimization: assessing changes of optimal orbitals for orthogonal multi-reference states." arXiv preprint arXiv:2211.08329 (2022) <https://doi.org/10.48550/arXiv.2211.08329>`_
