+++ paper.bib
@@ -164,4 +164,10 @@ Program title: QuTiP: The Quantum Toolbox in Python Catalog identifier: AEMB_v2_
 }
 
 
-
+@article{yalouz2022orthogonally,
+  title={Orthogonally Constrained Orbital Optimization: assessing changes of optimal orbitals for orthogonal multi-reference states},
+  author={Yalouz, Saad and Robert, Vincent},
+  journal={arXiv preprint arXiv:2211.08329},
+  year={2022},
+  url={https://doi.org/10.48550/arXiv.2211.08329}
+}
