### 2022-11-21 #4759 @SYalouz
Dear @jarvist, @wcwitt and @erikkjellgren,

For your information, I just updated the article with a brand new reference from our group mentioning the use of the QuantNBody package.

Also, as a quick reminder: all corrections and suggestions from @wcwitt and @erikkjellgren have been implemented.
We look forward to hearing your feedback on them!

Best regards
