### 2022-11-22 #4759 @erikkjellgren
Hi @SYalouz

Thank you for addressing all the points so nicely.
Here a few issues/suggestions I have left.

**Automated tests:**
It is very nice there is an automated test setup now.
A nice feature from a users perspective would be a meassurement of well tested to code is, this could be by lines of code hit by the tests, see tools like coveralls.io. (This is just a nice to-have suggestion).

**Community guidelines:**
I think the community guidelines should also be in the documentation.
Some users might never look at the README.md if they seek information.

**Quality of writing:**
Line 35: Openfermion should be OpenFermion

**Documentation:**
There seem to be a lot of types and that sort of stuff in the documentation.
Tho, it is not so bad that it makes the documentation confusing.
I am sure the documentation can just be cleaned up later on, and should not be a problem for this JOSS publication.
