### 2022-11-22 a129bdc
corrections paper + guidelines
