### 2022-11-22 #4759 @SYalouz
Dear @erikkjellgren,

Thanks a lot for your very quick reply, we are happy that the modifications we made satisfy you !

Following you last remarks, I have just updated the documentations to include the community guidelines. I also corrected the typo about OpenFermion that you noticed in the paper.

We thank you also for your good advices concerning the automated tests and the documentations. We keep that in mind for future modifications.

We now wait for the feedback from @wcwitt !

Kind regards,

Saad
