### 2022-11-22 43d49e1
moving guidelines to intro
