+++ docs/source/intro.rst
@@ -78,3 +78,13 @@ To generate these data, the package was used to code (from scratch) each many-bo
 .. note ::
 
   To illustrate how to use the package, several tutorials are shown in the section ''Tutorials and examples''.
+
+
+  How to contribute / Support
+  ______________________________
+
+  We'd love to accept your contributions and patches to QuantNBody. There are a few small guidelines you need to follow.
+
+  All submissions require review. We use GitHub pull requests for this purpose. Consult GitHub Help for more information on using pull requests. Furthermore, please make sure your new code comes with documentation.
+
+  If you are having issues, please let us know by posting the issue on our Github issue tracker.
