### 2022-12-09 #4759 @SYalouz
Dear @wcwitt  (and @jarvist)

Thanks for the message.

As we reach the end of the period you mentioned (5-9 December), I come back to you to know how is the review going on your side ?
