### 2022-12-09 SYalouz/QuantNBody#4 @wcwitt
Great - the new documentation looks good.
