### 2022-12-09 SYalouz/QuantNBody#5 @wcwitt
I would prefer some kind of continuous integration, but I suppose JOSS does not strictly require that.
