### 2022-12-09 #4759 @wcwitt
I've made another round of review, which involved:

* installing from scratch;
* running the tests;
* running the main tutorials;
* spot-checking the other examples (`Demo_CASCI_LiH.py`, `Demo_FCI_ringFermiHubbard.py`, `Demo_FCI_BoseHubbard.py`);
* verifying the issues I raised at the repo (linked above) are addressed;
* re-reading the paper.

Overall things look good, and I think you're nearly there.

The `Demo_FCI_ringFermiHubbard.py` could not complete within 30 minutes on my laptop, although it finished very quickly when I reduced the number of active electrons and molecular orbitals from 8 to 6. Is this expected?
