### 2022-12-09 e4133d6
Corrections FCI Hubbard ring
