### 2022-12-09 #4759 @SYalouz
Dear @wcwitt,

Thanks for this remark concerning the example Demo_FCI_ringFermiHubbard.py
We forgot to update the size of the system, indeed 8 electron in 8 orbital is already a lot
and  the Scipy eigensolver will take time if we want to access the full spectrum here.

I corrected this by reducing the size of the system to 6 electrons in 6 orbitals.
Now it should run in a few seconds.
