### 2022-12-09 #4759 @wcwitt
I raised that point partly because I remembered this line from the paper

> As a result, QuantNBody allows the realization of larger scale calculations (up to 8 electrons in 8 spatial molecular orbitals in a few seconds).

Can you reconcile the two?
