### 2022-12-09 #4759 @SYalouz
Dear @wcwitt,

I see your point. I corrected the code to use a sparse eigensolver from scipy.
Now you can produce the results of 8 electrons in 8 MOs in a few seconds.
