### 2022-12-09 746bee5
cleaning the printing tests
