### 2022-12-15 #4759 @SYalouz
Hi @jarvist !

Do you mean something like a zenodo DOI ?
