### 2022-12-15 #4759 @jarvist
Yes! A software archive (with DOI) that will act as a permanent source of the software as reviewed by JOSS. I find the easiest way is to generate a release within GitHub & then have Zenodo mint a DOI + archive from that. But we can also accept uploading a Zip file to figshare or similar.
