### 2022-12-15 #4759 @jarvist
Oh dear! That's one of the Guru meditations that might require an editorialbot developer to fix. Previously something similar was triggered by an error in the BibTex. I also see there's supposedly a missing DOI. Could you have a first look @SYalouz ?
