+++ paper.bib
@@ -154,13 +154,18 @@ Program title: QuTiP: The Quantum Toolbox in Python Catalog identifier: AEMB_v2_
   year={2022},
   url={https://doi.org/10.48550/arXiv.2210.02325}
 }
+ 
 
 @article{yalouz2022quantum,
-  title={Quantum embedding of multi-orbital fragments using the Block-Householder-transformation},
-  author={Yalouz, Saad and Sekaran, Sajanthan and Sauban{\`e}re, Matthieu},
-  journal={arXiv preprint arXiv:2209.10302},
+  title={Quantum embedding of multi-orbital fragments using the block-Householder transformation},
+  author={Yalouz, Saad and Sekaran, Sajanthan and Fromager, Emmanuel and Sauban{\`e}re, Matthieu},
+  journal={The Journal of Chemical Physics},
+  volume={157},
+  number={21},
+  pages={214112},
   year={2022},
-  url={https://doi.org/10.48550/arXiv.2209.10302}
+  publisher={AIP Publishing LLC},
+  url={https://doi.org/10.1063/5.0125683}
 }
 
 
