### 2022-12-15 #4759 @SYalouz
I just updated the reference of the article ;) It was an arxiv reference, but in the meantime the article has been accepted and published in a journal.
