### 2022-12-15 #4759 @jarvist
That might have been a bit fast for your changes to sync. Also, the error is referring to a figure. I wonder if it's double defined in your .TeX somehow.
