### 2022-12-15 #4759 @jarvist
Ah! You've given them both the same label name: `\label{fig:example}`. I think this should be unique.
