### 2022-12-15 #4759 @SYalouz
Yes, and actually I didn't use it in my text as I directly wrote figure and figure 2... Anyway, I have just erased this.
