+++ paper.bib
@@ -26,7 +26,7 @@
 
 
 @article{gray2018quimb,
-  title={quimb: a python library for quantum information and many-body calculations},
+  title={quimb: a {Python} library for quantum information and many-body calculations},
   author={Gray, Johnnie},
   journal={Journal of Open Source Software},
   year = {2018},
@@ -36,7 +36,7 @@
 }
 
 @article{weinberg2017quspin,
-  title={QuSpin: a Python package for dynamics and exact diagonalisation of quantum many body systems part I: spin chains},
+  title={QuSpin: a Python package for dynamics and exact diagonalisation of quantum many body systems. Part I: spin chains},
   author={Weinberg, Phillip and Bukov, Marin},
   journal={SciPost Physics},
   volume={2},
@@ -86,8 +86,8 @@
             Harris, Charles R. and Archibald, Anne M. and
             Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and
             {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
-  title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific
-            Computing in Python}},
+  title   = {{SciPy} 1.0: Fundamental Algorithms for Scientific
+            Computing in {Python}},
   journal = {Nature Methods},
   year    = {2020},
   volume  = {17},
@@ -97,11 +97,12 @@
 }
 
 @inproceedings{lam2015numba,
-  title={Numba: A llvm-based python jit compiler},
+  title={Numba: A {LLVM}-based Python {JIT} compiler},
   author={Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
   booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
   pages={1--6},
-  year={2015}
+  year={2015},
+  doi={10.1145/2833157.2833162}
 }
 
 
@@ -136,7 +137,7 @@ Program title: QuTiP: The Quantum Toolbox in Python Catalog identifier: AEMB_v2_
 }
 
 @article{luo2020yao,
-  title={Yao. jl: Extensible, efficient framework for quantum algorithm design},
+  title={Yao.jl: Extensible, efficient framework for quantum algorithm design},
   author={Luo, Xiu-Zhe and Liu, Jin-Guo and Zhang, Pan and Wang, Lei},
   journal={Quantum},
   volume={4},
