### 2022-12-15 #4759 @kyleniemeyer
Hi @SYalouz, can you please modify the metadata of the Zenodo archive such that the second author's name matches the paper? (I think currently it is using the GitHub handle.)

Also, I made a few typographic fixes to the paper and bibliography, can you merge the PR? https://github.com/SYalouz/QuantNBody/pull/6
