### 2022-12-15 #4759 @SYalouz
AMAZING, the DOI works now  !

Thanks a lot everyone :D

Next package I develop,  I will not hesitate to come back to JOSS !!!!
